version 1.5.9 (#75)
* refactor: Remove useless code and add log

* feat: Make init async ans throwable

Wait until open the document and throw an error if it is not successful.

* refactor: Remove useless function


Since pdfDocument is initialized at init in viewModel, use it right away.

* refactor: Use Task

Since viewModel creation is async, await and show view on the main thread.

* refactor: Use Task

Since viewModel creation is async, await and show view on the main thread.

* refactor: Make document not optional

* fix: Make `openSuccess()` throwable

Catch error if pdfDocument is nil

* refactor: Make property not optional

* fix: Initialize with url, not VC

Before: memory leak happened
If open a file that's already open, there's no need to create a VC, so just take the URL.

* feat: Add property

* feat: Check url and make new document if needed

* refactor: Refactor

* feat: Add `UILabel+`

Add `setHeaderStyle()` and `setBodyStyle()`.

* doc|refactor: Add description and refactor

If bounds's width or height is 0, there's no need to add shadowLayer.

* feat: Add `divider()`

* feat: Add `UITapGestureRecognizer+`

Add `UITapGestureRecognizerWithClosure` that can make UITapGestureRecognizer with closure.

* feat: Add Constants

* feat: Add `RadioButtonShape`

A view with a small circle within a circle.

* feat: Add `RadioButton`

Can select or deselect.

* feat: Add `RadioButtons`

Object that has RadioButton array.
Controls the state.

* Feature/move strategy setting view (#70)

* feat: Add `AnimatableView`

Add protocol about the view that perform the animation.

* feat: Add `MoveStrategyType`

* feat: Add `MoveStrategyAnimationView`

Shows the moving method according to MoveStrategy as an animation.

* feat: Add `SettingViewStyle`

protocol about the standard settingView.

* feat: Add `ExpandableSettingViewStyle`

protocol about the expandable settingView. should conform `SettingViewStyle`.

* feat: Add `UserDefaultsHelper`

Can save moveStrategy with propertyWrapper `UserDefault`.

* feat: Add `MoveStrategySettingView`

Change moveStrategy and post notification.

* Feat/settings view (#71)

* fix|feat: Fix constraints error and improve design

In `fold()`, there was a constraints error because bottom and top were the same, so fix it.
After that, since the `content` can go out of the cell by transform, add a `contentView` that wraps it.

Improve the design by setting the background color of the `contentView`.

* feat: Add `SettingsView`

* feat: Set height

* feat|doc: Add `showSettings()`

* feat: Add `settingsButton` and action

* feat: Change moveStrategy by user setting

Can update strategy whenever the value changes in settings because subscribed notification.

* doc: Add descriptions

* Feat/play mode setting view (#72)

* design: Adjust colors and Add color

* feat: Add `playModeAutoScale`

* feat: Add `setAutoScale()`

Add `setAutoScale()` in DocumentViewDelegate.
When play mode starts, set autoScale as UserDefaults.

* feat: Add `Size`

Can choose `RadioButtonShape`'s size.

* feat: Add `TogglableButton`

A toggleable button using `RadioButtonShape`.

* feat: Add `SettingStackCell`

The view that defines the style of each cell used in the Setting's stackView.
There are two types: `.toggleable` and `.withIcon`.

* feat: Add `PlayModeSettingView`

Can set autoScale when start play mode.

* refactor: Add `UIImage` extension

* feat: Add `UIStackView` extension

Add function `addArrangedSubviewWithDivider()`.

* feat: Add play mode setting

* Feat/tap area (#73)

* feat: Add `LineViewStyle`

Defines the view to be used in the screen where the Tap Area is set.

* feat: Add `HorizontalLineView` and `VerticalLineView`

* feat: Add `AnimatableGradientBackground`

Can add gradient animation.

* feat: Add `setSeconderyHeaderStyle()`

* feat: Add `playModeTapAreaAxis` and `playModeTapAreaLength`

* feat: Add `UIPanGestureRecognizerWithClosure`

* feat: Add `TapAreaAxis`

Determine the area according to the TapAreaAxis and length and moves it.

* feat: Add `PlayModeTapAreaSettingView`

Can set tap area with dragging the view.

* feat: Add setting tap area

* feat: Make conform `PlayModeSettingDelegate`

* Fix/play button action (#74)

* feat: Make `playButtonAction()` return bool

* fix: Start play mode when return is true

Make it start only when the vm can start play mode.

* design|refactor: Set titleColor and use Configuration

* version update 1.5.9

Add settings.